+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - Add `lastEventId` method to the API ([#2675](https://github.com/getsentry/sentry-react-native/pull/2675))
 
+### Fix
+- `Sentry.startTransaction` doesn't require `op` ([#](https://github.com/getsentry/sentry-react-native/pull/))
+
 ### Dependencies
 
 - Bump Cocoa SDK from v7.31.2 to v7.31.3 ([#2647](https://github.com/getsentry/sentry-react-native/pull/2647))